backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
marshmallow==3.20.1
pandas
openpyxl==3.1.2
python-keycloak==3.9.0
werkzeug==3.0.1
